refactor(utils): split proto to a new file
